Moved discovery scripts again, this time to command-specific directory 'retrace'
